chore: use Stitches tokens instead of hardcoded values
Instead of using hardcoded values, use Stitches tokens for things like spaces,
padding, font-sizes, margins, etc.

This way the UI can hava a minimal set of sizes, and I think this way the
UI will look consisten.

It's not an easy thing to deal with, I need to learn more about this =)